Added first test
