+++ back-end/src/dms/tests/conftest.py
@@ -9,6 +9,7 @@ from django.contrib.auth import get_user_model
 from dms.models.documents import File
 
 from auth.models import User
+from lms.models.students import Student
 
 SUPERUSER_EMAIL = "superuserfortests@mail.com"
 SUPERUSER_PASSWORD = "superuserpasswordfortests"
@@ -117,6 +118,47 @@ def paper_data(file):
             "upload_date": upload_date,
             "publication_date": publication_date,
             "is_binned": is_binned
+        }
+
+    return call_me
+
+@pytest.fixture
+def student_data():
+
+    def call_me(
+        name: str = "first",
+        surname: str = "second",
+        patronymic: str = "patronymic",
+    ) -> tuple:
+
+        s = Student()
+        s.name = name
+        s.surname = surname
+        s.patronymic = patronymic
+        return s, {
+            "fullname": s.full_name,
+            "name": s.name,
+            "surname": s.surname,
+            "patronymic": s.patronymic,
+            "photo": s.photo,
+            "milgroup": s.milgroup,
+            "birth_info": s.birth_info,
+            "university_info": s.university_info,
+            "application_process": s.application_process,
+            "skills": [],
+            "contact_info": s.contact_info,
+            "citizenship": s.citizenship,
+            "permanent_address": s.permanent_address,
+            "surname_genitive": s.surname_genitive,
+            "name_genitive": s.name_genitive,
+            "patronymic_genitive": s.patronymic_genitive,
+            "status": s.status,
+            "post": s.post,
+            "user": s.user,
+            "milspecialty": s.milspecialty,
+            "passport": s.passport,
+            "recruitment_office": s.recruitment_office,
+            "family": []
         }
 
     return call_me
